Restructure HTTP1-specific code from SmokeOperations package to a new SmokeOperationsHTTP1 to provide a clear separation between protocol-sepcific and protocol-agnostic code.
